Implement multiple testing correction strategy for power simulations
Add option to use p-value threshold corresponding to 5% FDR in real differential expression test results to calculate significance in simulated results for power calculations. This treats significance of simulated enhancer effect as if a given simulated effect would have been detected if part of the real experiment. The alterative, previously implemented, approach is to calculate FDR on simulated results, which can lead to different FDR behavior due to the much higher proportion of low p-values.